requirements.txt
gradio
matplotlib
pandas
numpy
nltk
wordcloud
pillow